requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
qdrant-client==1.11.3
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-qdrant==0.2.0
streamlit==1.41.1
gradio_client==1.4.3
python-dotenv==1.0.1
requests==2.32.3
Pillow==11.0.0
google-generativeai==0.8.3
